+++ application/index/controller/Ajax.php
@@ -470,180 +470,8 @@ class Ajax extends Base
         session($param['type'].'_verify','1');
         return json(['code'=>1,'msg'=>lang('ok')]);
     }
-    /**
-     *  获取分类树
-     *
-     * @param Request $request
-     * @return \think\response\Json
-     */
-    public function get_list(Request $request)
-    {
-        // 参数校验
-        // 参数校验
-        $param = $request->param();
-        $validate = validate($request->controller());
-        if (!$validate->scene($request->action())->check($param)) {
-            return json([
-                'code' => 1001,
-                'msg'  => '参数错误: ' . $validate->getError(),
-            ]);
-        }
-        // 查询条件组装
-        $where = [];
-        // 查询第一级
-        $where['type_pid'] = 0;
-
-        if (isset($param['type_id'])) {
-            $where['type_id'] = (int)$param['type_id'];
-        }
-
-        // 数据获取
-        $total = model('Type')->getCountByCond($where);
-        $list = [];
-        if ($total > 0) {
-            // 排序
-            $order = "type_sort DESC";
-            $field = '*';
-            $list = model('Type')->getListByCond(0, PHP_INT_MAX, $where, $order, $field, []);
-            foreach ($list as $index => $item) {
-                $child_total = Db::table('mac_type')->where(['type_pid' => $item['type_id']])->count();
-                if ($child_total > 0) {
-                    $child = Db::table('mac_type')->where(['type_pid' => $item['type_id']])->order('type_sort ASC')->select();
-                    $list[$index]['child'] = $child;
-                }
-            }
-        }
-        // 返回
-        return json([
-            'code' => 1,
-            'msg'  => '获取成功',
-            'info' => [
-                'total'  => $total,
-                'rows'   => $list,
-            ],
-        ]);
-    }
-
-    /**
-     *  获取视频列表
-     *
-     * @param Request $request
-     * @return \think\response\Json
-     */
-    public function get_vod_list(Request $request)
-    {
-        // 参数校验
-        $param = $request->param();
-        $validate = validate($request->controller());
-        if (!$validate->scene($request->action())->check($param)) {
-            return json([
-                'code' => 1001,
-                'msg'  => '参数错误: ' . $validate->getError(),
-            ]);
-        }
-        $offset = isset($param['offset']) ? (int)$param['offset'] : 0;
-        $limit = isset($param['limit']) ? (int)$param['limit'] : 20;
-        // 查询条件组装
-        $where = [];
-        if (isset($param['type_id'])) {
-            $where['type_id'] = (int)$param['type_id'];
-        }
-        if (isset($param['id'])) {
-            $where['vod_id'] =(int) $param['id'];
-        }
-//        if (isset($param['type_id_1'])) {
-//            $where['type_id_1'] = (int)$param['type_id_1'];
-//        }
-        if (!empty($param['vod_letter'])) {
-            $where['vod_letter'] = $param['vod_letter'];
-        }
-        if (isset($param['vod_tag']) && strlen($param['vod_tag']) > 0) {
-            $where['vod_tag'] = ['like', '%' . $this->format_sql_string($param['vod_tag']) . '%'];
-        }
-        if (isset($param['vod_name']) && strlen($param['vod_name']) > 0) {
-            $where['vod_name'] = ['like', '%'.$this->format_sql_string($param['vod_name']).'%'];
-        }
-        if (isset($param['vod_blurb']) && strlen($param['vod_blurb']) > 0) {
-            $where['vod_blurb'] = ['like', '%' . $this->format_sql_string($param['vod_blurb']) . '%'];
-        }
-        if (isset($param['vod_class']) && strlen($param['vod_class']) > 0) {
-            $where['vod_class'] = ['like', '%' . $this->format_sql_string($param['vod_class']) . '%'];
-        }
-        if (isset($param['vod_area']) && strlen($param['vod_area']) > 0) {
-            $where['vod_area'] = $this->format_sql_string($param['vod_area']);
-        }
-        if (isset($param['vod_year']) && strlen($param['vod_year']) > 0) {
-            $where['vod_year'] = $this->format_sql_string($param['vod_year']);
-        }
-        // 数据获取
-        $total = model('Vod')->getCountByCond($where);
-        $list = [];
-        if ($total > 0) {
-            // 排序
-            $order = "vod_time DESC";
-            if (strlen($param['orderby']) > 0) {
-                $order = 'vod_' . $param['orderby'] . " DESC";
-            }
-            $field = 'vod_id,vod_name,vod_actor,vod_hits,vod_hits_day,vod_hits_week,vod_hits_month,vod_time,vod_remarks,vod_score,vod_area,vod_year,vod_tag,vod_pic,vod_pic_thumb,vod_pic_slide,vod_douban_score';
-//            $list = model('Vod')->getListByCond($offset, $limit, $where, $order, $field, []);
-            $list = model('Vod')->getListByCond($offset, $limit, $where, $order, $field);
-            //把vod_time 字段转换为时间字符串
-            foreach ($list as &$value) {
-                $value['vod_time'] = date('Y-m-d H:i:s', $value['vod_time']);
-            }
-        }
-        // 返回
-        return json([
-            'code' => 1,
-            'msg'  => '获取成功',
-            'info' => [
-                'offset' => $offset,
-                'limit'  => $limit,
-                'total'  => $total,
-                'rows'   => $list,
-            ],
-        ]);
-    }
-    /**
-     * 视频详细信息
-     *
-     * @param Request $request
-     * @return \think\response\Json
-     * @throws \think\db\exception\DataNotFoundException
-     * @throws \think\db\exception\ModelNotFoundException
-     * @throws \think\exception\DbException
-     */
-    public function get_vod_detail(Request $request)
-    {
-        $param = $request->param();
-        $validate = validate($request->controller());
-        if (!$validate->scene($request->action())->check($param)) {
-            return json([
-                'code' => 1001,
-                'msg'  => '参数错误: ' . $validate->getError(),
-            ]);
-        }
-
-        $res = Db::table('mac_vod')->where(['vod_id' => $param['vod_id']])->find();
-        //判断vod_rel_vod 字段是否为空
-        if (!empty($res['vod_rel_vod'])) {
-            $field = 'vod_id,vod_name,vod_actor,vod_hits,vod_hits_day,vod_hits_week,vod_hits_month,vod_time,vod_remarks,vod_score,vod_area,vod_year,vod_tag,vod_pic,vod_pic_thumb,vod_pic_slide,vod_douban_score';
-            $res['vod_rel_vod_list'] = Db::table('mac_vod')->where(['vod_id' => ['in', $res['vod_rel_vod']]])->field($field)->select();
-        }
-        // 返回
-        return json([
-            'code' => 1,
-            'msg'  => '获取成功',
-            'info' => $res
-        ]);
-    }
-    protected function format_sql_string($str)
-    {
-        $str = preg_replace('/\b(SELECT|INSERT|UPDATE|DELETE|DROP|UNION|WHERE|FROM|JOIN|INTO|VALUES|SET|AND|OR|NOT|EXISTS|HAVING|GROUP BY|ORDER BY|LIMIT|OFFSET)\b/i', '', $str);
-        $str = preg_replace('/[^\w\s\-\.]/', '', $str);
-        $str = trim(preg_replace('/\s+/', ' ', $str));
-        return $str;
-    }
+    
+    
 
     /**
      * 获取APP图片列表（启动页、Banner广告）
